requirements.txt
SpeechRecognition
pyaudio
python-dotenv
deepl